+++ src/main/java/net/mohron/skyclaims/command/argument/IslandArgument.java
@@ -19,80 +19,65 @@
 package net.mohron.skyclaims.command.argument;
 
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
+import com.google.common.collect.Sets;
 import net.mohron.skyclaims.SkyClaims;
-import net.mohron.skyclaims.config.type.PermissionConfig;
 import net.mohron.skyclaims.permissions.Permissions;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.ArgumentParseException;
 import org.spongepowered.api.command.args.CommandArgs;
 import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.args.CommandElement;
+import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.format.TextColors;
 
 import javax.annotation.Nullable;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
-public class TargetArgument extends CommandElement {
+public class IslandArgument extends CommandElement {
 	private static final SkyClaims PLUGIN = SkyClaims.getInstance();
-	private static PermissionConfig config = PLUGIN.getConfig().getPermissionConfig();
 
-	private static final Map<String, Targets> TARGETS = Maps.newHashMap();
-
-	static {
-		TARGETS.put("island", Targets.ISLAND);
-		TARGETS.put("i", Targets.ISLAND);
-		TARGETS.put("chunk", Targets.CHUNK);
-		TARGETS.put("c", Targets.CHUNK);
-		TARGETS.put("block", Targets.BLOCK);
-		TARGETS.put("b", Targets.BLOCK);
-	}
-
-	public TargetArgument(@Nullable Text key) {
+	public IslandArgument(@Nullable Text key) {
 		super(key);
 	}
 
 	@Nullable
 	@Override
 	protected Object parseValue(CommandSource source, CommandArgs args) throws ArgumentParseException {
-		String target = args.next().toLowerCase();
-		if (TARGETS.containsKey(target)) {
-			if (!hasPermission(source, TARGETS.get(target)))
-				throw new ArgumentParseException(Text.of(TextColors.RED, "You do not have permission to use the supplied target!"), target, 0);
-			return TARGETS.get(target);
+		String arg = args.next().toLowerCase();
+		if (SkyClaims.islands.isEmpty())
+			throw new ArgumentParseException(Text.of(TextColors.RED, "There are no valid island!"), arg, 0);
+		try {
+			UUID uuid = UUID.fromString(arg);
+			if (SkyClaims.islands.containsKey(uuid)) {
+				Set<UUID> islands = Sets.newHashSet();
+				islands.add(uuid);
+				return islands;
+			}
+		} catch (IllegalArgumentException ignored) {
 		}
-		throw new ArgumentParseException(Text.of(TextColors.RED, "Invalid target!"), target, 0);
+		return SkyClaims.islands.entrySet().stream()
+				.filter(i -> i.getValue().getOwnerName().equalsIgnoreCase(arg))
+				.map(Map.Entry::getKey)
+				.collect(Collectors.toSet());
 	}
 
 	@Override
 	public List<String> complete(CommandSource src, CommandArgs args, CommandContext context) {
 		try {
-			String name = args.peek().toLowerCase();
-			return TARGETS.entrySet().stream()
-					.filter(s -> s.getKey().length() > 1)
-					.filter(s -> s.getKey().startsWith(name))
-					.filter(s -> hasPermission(src, s.getValue()))
-					.map(Map.Entry::getKey)
+			String arg = args.peek().toLowerCase();
+			boolean admin = src.hasPermission(Permissions.COMMAND_LIST_ALL);
+			return SkyClaims.islands.entrySet().stream()
+					.filter(i -> i.getValue().getOwnerName().toLowerCase().startsWith(arg))
+					.filter(i -> admin || !(src instanceof Player) || i.getValue().hasPermissions((Player) src))
+					.map(m -> m.getValue().getOwnerName())
 					.collect(Collectors.toList());
 		} catch (ArgumentParseException e) {
 			return Lists.newArrayList();
-		}
-	}
-
-	private boolean hasPermission(CommandSource src, Targets target) {
-		if (!config.isSeparateTargetPerms()) return true;
-		switch (target) {
-			case BLOCK:
-				return src.hasPermission(Permissions.COMMAND_ARGUMENTS_BLOCK);
-			case CHUNK:
-				return src.hasPermission(Permissions.COMMAND_ARGUMENTS_CHUNK);
-			case ISLAND:
-				return true;
-			default:
-				return false;
 		}
 	}
 }